refactor: enhanced user endpoints, added custom classes for data validation, added birth_date to user schema
